@remotion/studio: Extend timeline unselection area
