+++ packages/studio/src/components/Timeline/TimelineList.tsx
@@ -1,8 +1,7 @@
-import React, {useCallback} from 'react';
+import React from 'react';
 import {BACKGROUND} from '../../helpers/colors';
 import type {TrackWithHash} from '../../helpers/get-timeline-sequence-sort-key';
 import {TimelineListItem} from './TimelineListItem';
-import {useTimelineSelection} from './TimelineSelection';
 import {TimelineTimePadding} from './TimelineTimeIndicators';
 
 const container: React.CSSProperties = {
@@ -13,24 +12,8 @@ const container: React.CSSProperties = {
 export const TimelineList: React.FC<{
 	readonly timeline: TrackWithHash[];
 }> = ({timeline}) => {
-	const {clearSelection} = useTimelineSelection();
-
-	// Selection-triggering click handlers in children call e.stopPropagation(),
-	// so any pointerdown that bubbles up here is by definition on empty space
-	// and should clear the current selection.
-	const onPointerDown = useCallback(
-		(e: React.PointerEvent<HTMLDivElement>) => {
-			if (e.button !== 0) {
-				return;
-			}
-
-			clearSelection();
-		},
-		[clearSelection],
-	);
-
 	return (
-		<div style={container} onPointerDown={onPointerDown}>
+		<div style={container}>
 			<TimelineTimePadding />
 			{timeline.map((track) => {
 				return (
